A star algorithm script, new Virus script and portal, script
The a star algo was finished and the virus script was developed to allow the viruses to spawn on the portal at the beginning of the game.  The viruses can now move from the start point to the end point independently while avoiding obstacles such as towers or the edge of the map.  It will find the shortest path.